+++ recipes/validate.rb
@@ -104,5 +104,8 @@ end
 if is_certificate_server
   include_recipe 'is_apaas_openshift_cookbook::adhoc_migrate_certificate_server' if helper.check_certificate_server
 end
-include_recipe 'is_apaas_openshift_cookbook::adhoc_redeploy_certificates' if node['is_apaas_openshift_cookbook']['adhoc_redeploy_certificates']
-include_recipe 'is_apaas_openshift_cookbook::commons' unless node.run_state['issues_detected']
+
+unless node['is_apaas_openshift_cookbook']['upgrade']
+  include_recipe 'is_apaas_openshift_cookbook::adhoc_redeploy_certificates' if node['is_apaas_openshift_cookbook']['adhoc_redeploy_certificates']
+  include_recipe 'is_apaas_openshift_cookbook::commons' unless node.run_state['issues_detected']
+end